Address review: add text-dark to badge and extract extension
- Add text-dark class to bg-info badge for better contrast (accessibility)
- Move WorkshopLanguage.localizedName(for:) to dedicated file for reuse
